Make PointGrid into a compute method
